Move RemoveIrrelevantField to pkg/resourceinterpreter
